Moved to 1.0.0 - Fixed #76 : renamed endpoints to retrievers for clarity
