+++ MyCapytain/__init__.py
@@ -10,5 +10,5 @@
 """
 
 __name__ = "MyCapytain"
-__version__ = "0.1.3"
-__all__ = ["common", "endpoints", "resources"]
+__version__ = "1.0.0"
+__all__ = ["common", "retrievers", "resources"]
