--- MyCapytain/endpoints/__init__.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-.. module:: MyCapytain.endpoints
-   :synopsis: Implementation and abstraction of CTS related endpoints
-
-.. moduleauthor:: Thibault Clérice <leponteineptique@gmail.com>
-
-
-"""
