+++ MyCapytain/retrievers/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""
+.. module:: MyCapytain.retrievers
+   :synopsis: Implementation and abstraction of CTS related retrievers
+
+.. moduleauthor:: Thibault Clérice <leponteineptique@gmail.com>
+
+
+"""
