+++ README.rst
@@ -12,6 +12,8 @@
    :target: https://zenodo.org/badge/latestdoi/3923/Capitains/MyCapytain
 .. image:: https://img.shields.io/pypi/dm/MyCapytain.svg
    :target: https://pypi.python.org/pypi/MyCapytain
+.. image:: https://api.codacy.com/project/badge/grade/8e63e69a94274422865e4f275dbf08ea
+   :target: https://www.codacy.com/app/leponteineptique/MyCapytain
 MyCapytain is a python package which provides a large set of tools to deal with CTS and Capitains Guidelines in Python.
 
 Installation and Requirements
